backend/requirements.txt
browser_use_sdk
python-dotenv
fastapi
pydantic